requirements.txt
Django>=3.0,<4.0
psycopg2-binary==2.9.3
django-crispy-forms==1.12.0
djangorestframework==3.12.4
django-environ==0.11.2
python-dotenv==1.0.1
psycopg2==2.9.9
cffi==1.16.0
cryptography==42.0.2
pycparser==2.21
bcrypt==4.1.2
paramiko==3.4.0
pynacl==1.5.0
